refactor: 바우처 관련 메시지 인터페이스에서 관리하도록 변경
