+++ src/main/java/com/programmers/voucher/domain/voucher/controller/VoucherController.java
@@ -3,11 +3,14 @@ package com.programmers.voucher.domain.voucher.controller;
 import com.programmers.voucher.domain.voucher.domain.Voucher;
 import com.programmers.voucher.domain.voucher.dto.request.VoucherCreateRequest;
 import com.programmers.voucher.domain.voucher.service.VoucherService;
+import com.programmers.voucher.domain.voucher.util.VoucherMessages;
 import com.programmers.voucher.global.io.Console;
 import org.springframework.stereotype.Controller;
 
 import java.util.List;
 import java.util.UUID;
+
+import static com.programmers.voucher.domain.voucher.util.VoucherMessages.CREATED_NEW_VOUCHER;
 
 @Controller
 public class VoucherController {
@@ -23,7 +26,8 @@ public class VoucherController {
         VoucherCreateRequest request = console.inputVoucherCreateInfo();
         UUID voucherId = voucherService.createVoucher(request.getVoucherType(), request.getAmount());
 
-        console.print("Created new voucher. VoucherID: " + voucherId.toString());
+        String consoleMessage = VoucherMessages.addVoucherId(CREATED_NEW_VOUCHER, voucherId.toString());
+        console.print(consoleMessage);
     }
 
     public void findVouchers() {
